+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Next Release
 
+## 0.0.51
+
+- [contracts] Final updates from auditor
+- [contracts] Add price oracle contract
+- [router] Estimate gas and add to fee on configured chains
+- [contracts] Deploy on testnets
+
+## 0.0.50
+
+- [sdk] Hardcode estimated gas for xdai
+
 ## 0.0.49
 
 - [txservice] Properly handle insufficient funds errors